Combine test and train SUMO classes into a single SUMO environment class
Adds vehicle_stats as an optional parameter to the SUMO training class.
As this is the only distinction between the train and test classes, the test class is deleted.
Gets rid of the Environment folder, and instead renames the SUMO script to environment.
Updates usage of the SUMO class as well.