Add support to change any config option from CLI
Changes the settings to allow CLI options for any of the parameters in the config file.
CLI options for the parameter 'example_param' are of the form '--example-param', followed by the parameter's value.
Boolean parameters do not need a value after, and are of the form '--flag' to set the value to True, or '--not-flag' to set it to False.
List parameters must be a comma-delimited list without spaces in between values.