Improve action combinations computation for non-fixed action space
Changes how action combinations are generated when the action space is not fixed.
Instead of generating all possible action combinations in a list, and then using an actions index to retrieve the combination that specifies the action for every intersection, the action combination is now

The action combination can be viewed as converting the actions index (in base 10) to a number in base N, where N is the number of possible actions, and each digit of this result is then the action of a single intersection.
This result must also be padded with leading zeros so that the number of digits is equal to the number of intersections.