Add noisy net support to Rainbow DQN agent
Adds a new NoisyLinear layer class and NoisyNet class to implement an alternative to epsilon-greedy exploration.
Adds a noisy_net parameter to Rainbow DQN to toggle it.
Refactors the existing Net class, so that it is a subclass of a new AbstractNet class. This allows Net and NoisyNet to be simple subclasses of AbstractNet with the only main difference being the use of Linear or NoisyLinear layers.
Also cleans up the model module and removes unused network classes.