Refactor model module to use hierarchy of network classes
Refactors the neural network classes to use inheritance to reduce code duplication.
Creates AbstractNet class to be used as a model superclass.
Separates the ActorNet and CriticNet classes each into new distinct classes for the PPO and WOLP algorithms.
Rename Net to QNet.
Adjusts the usage of these network classes in all agents.