Add support for enum training settings
Adds support in the settings module for training settings that are enums (a limited set of string values).
In the config file, the enum values can be written case insensitively.
Adds enums for the training_strategy, reward_definition, state_representation and agent_type settings.
Simplifies the way to specify a type for a setting (instead of a string for the type, just uses the type itself. For example, `'gui': bool` instead of `'gui': 'bool'`).