Refactor memory module and N-step bootstrapping
Creates a new memory directory.
Separates the previous memory module into a replay module (for experience replay classes) and a structures module (for necessary data structures) in the new directory.
Refactors NStepReplayBuffer to take a replay buffer instance as a parameter, and use this as a backing buffer to then act as a replay buffer itself. This means it can easily integrate with a normal replay buffer or a prioritized replay buffer.