Add continuous action space support to PPO
Adds an action_definition parameter to PPO that specifies whether to  use phase/cycle actions (discrete or continuous action space).

Adds a min_green_duration to the training settings, that specifies the minimum duration of a green light phase in seconds for the continuous action space.

Adds a _compute_total_green_time method to SumoCycleEnv, to compute how much green light time is available during a cycle to be distributed amongst the different green light phases, taking into account the fixed yellow and red light phases, and the minimum green light duration.

Adds PpoContinuousActorNet and PpoContinuousCriticNet network classes to the model module.
Adds a new TanhNet that uses the Tanh activation function to support these new network classes.
Refactors the ReluNet and TanhNet classes to inherit from a base DenseNet class to reduce code duplication.

Adds support to the PPO agent and environment for a continuous action space, where actions are a partitioning of the available green light time to all traffic light phases. This is achieved as follows:
    - A continuous actor network learns the alpha and beta parameters for several Beta distributions (one distribution for each traffic light phase).
    - For each traffic phase, a value (in the interval [0, 1]) is sampled from its corresponding Beta distribution.
    - These values represent the actions outputted by the agent. These raw actions are then transformed into meaningful actions in the environment with the _compute_cycle_action function in SumoCycleEnv.
    - This involves passing all raw actions through a Softmax function, to give each traffic phase a proportion value, where all values sum to 1.
    - Each proportion value is multiplied by the available green light time, and then added to the minimum green light time, to give the traffic green phase in seconds.
    - This ensures that all traffic phase durations sum up to the green light time (and thus all phases sum to the cycle time).

Also some minor refactoring of the PPO agent and environment module.